src/requirements.txt
discord.py
requests
beautifulsoup4
pandas
python-telegram-bot
redis